# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Sarathi API
  version: 0.1.0
paths:
...
  /api/v1/lessons/{id}:
    get:
      tags:
      - lessons
      summary: Get Lesson Metadata
      description: 17 GET /api/v1/lessons/{id} - metadata only
      operationId: get_lesson_metadata_api_v1_lessons__id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lessons/{id}/start:
    post:
      tags:
      - lessons
      summary: Start Lesson
      description: 18 POST /api/v1/lessons/{id}/start - marks in progress, returns
        thread id
      operationId: start_lesson_api_v1_lessons__id__start_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lessons/{id}/content:
    get:
      tags:
      - lessons
      summary: Stream Lesson Content
      description: 19 GET /api/v1/lessons/{id}/content - SSE stream of ContentBlocks
      operationId: stream_lesson_content_api_v1_lessons__id__content_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lessons/{id}/reexplain:
    post:
      tags:
      - lessons
      summary: Reexplain Lesson Block
      description: 19b POST /api/v1/lessons/{id}/reexplain - SSE stream
      operationId: reexplain_lesson_block_api_v1_lessons__id__reexplain_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReexplainRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lessons/{id}/complete:
    post:
      tags:
      - lessons
      summary: Complete Lesson
      description: 20 POST /api/v1/lessons/{id}/complete
      operationId: complete_lesson_api_v1_lessons__id__complete_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lessons/{id}/signals:
    post:
      tags:
      - lessons
      summary: Create Lesson Signal
      description: 25 POST /api/v1/lessons/{id}/signals
      operationId: create_lesson_signal_api_v1_lessons__id__signals_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignalCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/users/me:
    get:
      tags:
      - me
      summary: Get Me
      description: 'Get the current learner''s profile.

        Lazily creates a new profile record if this is the learner''s first request.'
      operationId: get_me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
      security:
      - HTTPBearer: []
  /api/v1/users/me/signals:
    get:
      tags:
      - me
      summary: Get My Signals
      description: 26 GET /api/v1/users/me/signals
      operationId: get_my_signals_api_v1_users_me_signals_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SignalResponse'
                type: array
                title: Response Get My Signals Api V1 Users Me Signals Get
      security:
      - HTTPBearer: []
  /api/v1/users/me/progress:
    get:
      tags:
      - me
      summary: Get My Progress
      description: 27 GET /api/v1/users/me/progress
      operationId: get_my_progress_api_v1_users_me_progress_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MasteryStateResponse'
                type: array
                title: Response Get My Progress Api V1 Users Me Progress Get
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MasteryStateResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        concept_id:
          type: string
          format: uuid
          title: Concept Id
        score:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Score
        confidence:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Confidence
        attempts:
          type: integer
          title: Attempts
        created_at:
          type: string
          format: date-time
          title: Created At
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
      type: object
      required:
      - id
      - user_id
      - concept_id
      - score
      - confidence
      - attempts
      - created_at
      - last_seen_at
      title: MasteryStateResponse
    MeResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        locale:
          type: string
          title: Locale
        created_at:
          type: string
          format: date-time
          title: Created At
        has_learner_profile:
          type: boolean
          title: Has Learner Profile
          default: false
        profile_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Profile Version
      type: object
      required:
      - id
      - email
      - display_name
      - locale
      - created_at
      title: MeResponse
...
    ReexplainRequest:
      properties:
        block_id:
          type: string
          format: uuid
          title: Block Id
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - block_id
      title: ReexplainRequest
...
    SignalCreate:
      properties:
        block_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Block Id
        type:
          $ref: '#/components/schemas/SignalType'
        value:
          additionalProperties: true
          type: object
          title: Value
          default: {}
      type: object
      required:
      - type
      title: SignalCreate
    SignalResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        lesson_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Lesson Id
        block_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Block Id
        type:
          $ref: '#/components/schemas/SignalType'
        value:
          additionalProperties: true
          type: object
          title: Value
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - user_id
      - lesson_id
      - block_id
      - type
      - value
      - created_at
      title: SignalResponse
    SignalType:
      type: string
      enum:
      - checkpoint_score
      - confusion_flag
      - time_on_block
      - hint_requested
      - retry
      - inline_check_failed
      - skip
      - session_abandon
      - revisit
      title: SignalType
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer